server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pycryptodome==3.20.0
aiofiles==23.2.1
python-dotenv==1.0.0
